requirements.txt
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.112.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.0
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
pendulum==3.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
time-machine==2.15.0
tornado==6.4.1
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
